% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrp_levels.R
\name{calc_5essentials_points}
\alias{calc_5essentials_points}
\title{Calculate 5Essentials points}
\usage{
calc_5essentials_points(rating)
}
\arguments{
\item{rating}{5Esstianls rating as character vectors:}
}
\value{
a single element numeric vector
}
\description{
\code{calc_5essentials_points} calculates  points given 
5Essentials Rating
}
\details{
`rating` must be a characteer vector with one of the following:
\itemize{
\item "WO" = Well Organized
\item "O" = Organized
\item "MO" = Mostly Organized
\item "PO" = Partially Organized
\item "NYO" = Not Yet Organized
}
}

