% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrp_levels.R
\name{calc_access_points}
\alias{calc_access_points}
\title{Calculate Access percentage points}
\usage{
calc_access_points(pct)
}
\arguments{
\item{pct}{percentage of students making sufficient progress on Access}
}
\value{
a single element numeric vector
}
\description{
\code{calc_access_points} calculates  points for sufficient 
growth on Access
}

