% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sqrp_levels.R
\name{calc_ada_points}
\alias{calc_ada_points}
\title{Calculate ADA points}
\usage{
calc_ada_points(ada)
}
\arguments{
\item{ada}{ADA points}
}
\value{
a single element numeric vector
}
\description{
\code{calc_ada_points} calculates ADA points
}

