% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrp_levels.R
\name{calc_attainment_points}
\alias{calc_attainment_points}
\title{Calculate priority attainment percentile points}
\usage{
calc_attainment_points(pct)
}
\arguments{
\item{pct}{attainment percentile}
}
\value{
a single element numeric vector
}
\description{
\code{calc_attainment_points} calculates  points for meeting
exceding typical groth
}
