% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sqrp_levels.R
\name{calc_dqi_points}
\alias{calc_dqi_points}
\title{Calculate DQI  points}
\usage{
calc_dqi_points(pct)
}
\arguments{
\item{pct}{percentage DQI score}
}
\value{
a single element numeric vector
}
\description{
\code{calc_dqi_points} calculates  points given DQI score
}

