% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sqrp_levels.R
\name{calc_growth_points}
\alias{calc_growth_points}
\title{Calculate growth percentile points}
\usage{
calc_growth_points(growth_pctl)
}
\arguments{
\item{growth_pctl}{growth percentile}
}
\value{
a single element numeric vector
}
\description{
\code{calc_growth_points} calculates growht points
}

