% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrp_levels.R
\name{calc_me_growth_norms_points}
\alias{calc_me_growth_norms_points}
\title{Calculate priority typical growth percentage points}
\usage{
calc_me_growth_norms_points(pct)
}
\arguments{
\item{pct}{percent meeting/exceeding growth}
}
\value{
a single element numeric vector
}
\description{
\code{calc_me_growth_norms_points} calculates  points for meeting 
exceding typical groth
}

