% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrp_levels.R
\name{calc_pct_me}
\alias{calc_pct_me}
\title{Calculate percent meets/exceeds typical growth}
\usage{
calc_pct_me(growth)
}
\arguments{
\item{growth}{an `sqrp_growth` object}
}
\value{
a data frame
}
\description{
\code{calc_pct_me} calculates  typical growth percentage from 
an `sqrp_growth` object
}

