% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrp_levels.R
\name{calc_priority_growth_points}
\alias{calc_priority_growth_points}
\title{Calculate priority group growth percentile points}
\usage{
calc_priority_growth_points(growth_pctl)
}
\arguments{
\item{growth_pctl}{growth percentile}
}
\value{
a single element numeric vector
}
\description{
\code{calc_priority_growth_points} calculates growht points
}
