% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/school_growth_percentile.R
\name{collapse_grade_to_school}
\alias{collapse_grade_to_school}
\title{Calculates school-level (i.e, cohort) growth percentiles a la CPS's SQRP}
\usage{
collapse_grade_to_school(.data)
}
\arguments{
\item{.data}{data passed to function}
}
\value{
data.frame
}
\description{
\code{collapse_grade_to_school} calculates school level growth
using CPS specific rules (e.g., using a CPS district wide standard deviation for
schools that have more than one grade leve).
}

