% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_equate.R
\name{cps_equate}
\alias{cps_equate}
\title{Create vector of imputed prior spring scores}
\usage{
cps_equate(rit_score, subject, grade_level,
  cps_constants = cps_constants_2015)
}
\arguments{
\item{rit_score}{a vector of fall RIT scores}

\item{subject}{list a character vector of subjects. The only acceptable values are
"Mathematics" or "Reading".}

\item{grade_level}{A numeric or integer vector of grade levels.  Grade levels
below 3rd are ignored and the original RIT score is returned wihtout any
imputation/equating performed.}
}
\value{
a character vector of \code{length(rit_score)}.
}
\description{
\code{cps_equate} returns a character vector of prior sprng scores that are imputed
from fall scores.  This function uses the regression coefficients that described
on page 2 of the \emph{CPS Metric Guide for National School Growth Percentile (NWEA)}.
}
\details{
From the handbook: "[f]or students who do not have a valid spring pretest
score but did test in the fall, thier fall scores will be equated to an
equivalent spring score. This equated score will then be incorporated into
the school's spring average for the purpose  of calculating the
spring-to-spring growth percentile" (SQRP Handbook 12/5/2014, 25).

The regression coefficients are taken from page 2  of the \emph{CPS Metric Guide for National
School Growth Percentile (NWEA)}.  Equating is only performed for grades 3-8.
}
\examples{
cps_equate(175, "Reading", 6)
cps_equate(c(190, 175), c("Mathematics", "Reading"), c(6,7))
}

