% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensurer_contracts.R
\name{ensure_fall_data}
\alias{ensure_fall_data}
\title{Ensure fall data has proper columns}
\usage{
ensure_fall_data(.)
}
\arguments{
\item{.}{the thing to be ensured}
}
\description{
checks `names(.)` for correct columns
}

