% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensurer_contracts.R
\name{ensure_one_seasons_only}
\alias{ensure_one_seasons_only}
\title{Ensure only two test terms in data}
\usage{
ensure_one_seasons_only(.)
}
\arguments{
\item{.}{the thing to be ensured}
}
\description{
checks if `testterm` field has exactly two terms
}

