% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensurer_contracts.R
\name{ensure_subjects}
\alias{ensure_subjects}
\title{Ensure subjects only reading and mathematics}
\usage{
ensure_subjects(.)
}
\arguments{
\item{.}{the thing to be ensured}
}
\description{
checks if `measurementscale` field only contains "Mathematics" or
"Reading"
}

