% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_growth_percentile.R
\name{equate_fall_to_spring}
\alias{equate_fall_to_spring}
\title{Applies CPS fall-to-prior-spring equating}
\usage{
equate_fall_to_spring(growth_data = map_matched, fall_data = NA)
}
\arguments{
\item{growth_data}{season matched growth data frame}

\item{fall_data}{a data frame of fall data with columns
`studentid`, `measurmentscale`, `grade`, and `testritscore`}
}
\value{
data.frame with `nrow(growth_data)` rows and
`ncol(growth_data) + 1` columns
}
\description{
\code{equate_fall_to_spring} applies the function
\code{cps_equate} on fall data that must be passed in and returns
a data frame with missing prior_spring test scores updated with
equated scores.  An `equated` indcator is added to the
resulting data frame.
}
