% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrp_levels.R
\name{get_sqrp_level}
\alias{get_sqrp_level}
\title{Calculate sqrp level given sqrp points}
\usage{
get_sqrp_level(sqrp_points)
}
\arguments{
\item{sqrp_points}{sqpr points}
}
\value{
a single element factor vector
}
\description{
\code{get_sqrp_level} calculates SQRP level given SQRP points.
}

