% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/school_growth_percentile.R
\name{get_test_percentile}
\alias{get_test_percentile}
\title{Finds student national percentile rank}
\usage{
get_test_percentile(measurementscale, season, grade_level, ritscore)
}
\arguments{
\item{measurementscale}{the test subject.}

\item{season}{the test season, which is one of "Fall", "Winter", or "Spring".}

\item{grade_level}{the grade level at the time of the assessment}

\item{ritscore}{the test RIT score for which you want to know the test percentile.}
}
\value{
an integer vector
}
\description{
\code{get_test_percentile} looks up a student's national
percentile rank from \code{\student_status_norms_2011_dense_extended}.
}

