% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/priority_group.R
\name{priority_group}
\alias{priority_group}
\title{Recalculate school growth or attainment precentiles for
priority groups.}
\usage{
priority_group(sqrp_obj, group_column, group_id, ...)
}
\arguments{
\item{sqrp_obj}{either an \code{sqrp_growth} or \code{sqrp_attainment}
object create by running #' \code{\link{school_growth_percentile}} or
\code{\link{school_growth_percentile}}, respectivley, on NWEA MAP data.}

\item{group_column}{a single element character vector with the
name of the column in the original data that identifies priority
group membership.}

\item{group_id}{a single element character vector of the code in
\code{grou_column} that identifites whether ro not a student is
a member of the priority group of interest}

\item{...}{other arguments passed to either
\code{\link{school_growth_percentile}} or
\code{\link{school_growth_percentile}}}
}
\value{
an \code{sqrp_growth} and \code{priority_group_growth}
or \code{sqrp_attainment} and \code{priority_group_attain}
}
\description{
\code{priority_group} recalculates SQRP school growth
or attainment percentiles for priority groups for an \code{sqrp_growth}
or \code{sqrp_attainment} attainment objects.  The function subsets
the original data by the priority group and then reruns
\code{\link{school_growth_percentile}} or
\code{\link{school_growth_percentile}}, depending on the class of
\code{sqrp_obj}
}

