% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecting.R
\name{project_s2s}
\alias{project_s2s}
\title{Create cdf with projected growth from a single season cdf}
\usage{
project_s2s(.data, student_column = "studentid", grade_column = "grade",
  subject_column = "measurementscale", rit_column = "testritscore",
  term_column = "termname", dl_indicator = "iep", ell_indicator = "iep",
  school_indicator = "schoolname", fall_equate_scores = NULL,
  percent_cr = 0.1)
}
\arguments{
\item{.data}{a long format (i.e., combined) Comprehensive Data File}

\item{student_column}{column that uniquely identifies students with
respect to subject and term}

\item{grade_column}{column that identifies grade levels}

\item{subject_column}{which column identifies measurement scale}

\item{rit_column}{column that identifes RIT score}

\item{term_column}{identifies term name}

\item{dl_indicator}{identifies diverse learner status}

\item{ell_indicator}{identifies English Language Learner status}

\item{school_indicator}{column that identifies school}

\item{fall_equate_scores}{either a data.frame of fall MAP results with columns `studentid`,
`measurementscale`, `testritscore`, and `grade`, or `TRUE` if your equateding all fall scores
from fall-to-spring map data.  If a data.frame is passed, then these data are used to equate
missing pre-test (i.e., prior-spring) RIT scores per CPS guidelines. IF `TRUE` is passed, then all fall scores
are equated to the prior spring}

\item{percent_cr}{percent of students who make excatly KIPP Tiered
college ready growth.  This parameter is a scalar between in (0,1)
and is passed to \code{\link[stats]{rbinom}}'s \code{prob} paramter.}
}
\value{
list with four data frames attached
showing student-, grade-, and school-level growth percentiles
as well as the original data.
}
\description{
\code{project_s2s} accepts a single season, which
must be "Spring", cdf and creates  a second season with projected
growth, and appends the projected growth to the original cdf.
}
\details{
This function essentially creates a two season CDF where
the second, later season has anticipated results.  The resulting
CDF is suitable for passing to \code{\link{school_growth_percentile}}
and after filtering to the second season to \code{\link{school_attainment_percentile}}.

 The \code{percent_cr} is used to control the minimum pernctage
 of students in the CDF that make KIPP Tiered college ready
 growth. Students not picked to make CR ready growth have growth that
 drawn from \code{\link[stats]{rnorm}} with the mean and sd set by
 for each student the 2011 NWEA MAP student-level norms.

 Since CR students are identified \code{\link[stats]{rbinom}} and
 non CR growth is determeind by \code{\link[stats]{rnorm}} these
 project growth of of students is fundametnally stochastic.
 Consequently, this function is best suited for simulating
 growth.

 As with \code{\link{school_growth_percentile}}, you can pass single
 season CDF of fall results to impute prior spring scores for students
}
