% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/school_attainment_percentile.R
\name{school_attainment_percentile}
\alias{school_attainment_percentile}
\title{Calculate school growth percentiles}
\usage{
school_attainment_percentile(.data, student_column = "studentid",
  grade_column = "grade", subject_column = "measurementscale",
  rit_column = "testritscore", term_column = "termname",
  dl_indicator = "iep", ell_indicator = "ell", race_indicator = NULL,
  school_indicator = "schoolname")
}
\arguments{
\item{.data}{a long format (i.e., combined) Comprehensive Data File with the
second year (post_test) MAP results.}

\item{student_column}{column that uniquely identifies students with
respect to subject and term}

\item{grade_column}{column that identifies grade levels}

\item{subject_column}{which column identifies measurement scale}

\item{rit_column}{column that identifes RIT score}

\item{term_column}{identifies term name}

\item{dl_indicator}{identifies diverse learner status}

\item{ell_indicator}{identifies English Language Learner status}

\item{race_indicator}{identifies race and ethnicity data}

\item{school_indicator}{column that identifies school}
}
\value{
list with four data frames attached
}
\description{
\code{school_attainment_percentile} is a workhourse function
that calculates school attainment percentiles
}

