% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecting.R
\name{simulate_once_sqrp_data}
\alias{simulate_once_sqrp_data}
\title{Produce one iteration of SQRP simulation}
\usage{
simulate_once_sqrp_data(spring_pretest_cdf, fall_equate_cdf = NULL,
  school = "Ascend", pct_cr, aa = list(aa_col = "studentethnicgroup", aa_ind
  = "Black or African American"), iep = list(iep_col = "sped", iep_ind =
  TRUE), ...)
}
\arguments{
\item{spring_pretest_cdf}{single season pret-test, spring cdf}

\item{fall_equate_cdf}{single season fall cdf used to equate missing scores}

\item{school}{character providing a part of the name of the school in
\code{spring_test_cdf} for which to provide simulations,}

\item{pct_cr}{percent of students that should make exactly KIPP
Tiered college ready growth}

\item{aa}{a list with two elements: \code{aa_col} and \code{aa_ind},
which are use to id columns for priority group calculations}

\item{iep}{a list with two elements: \code{iep_col} and \code{iep_ind},
which are use to id columns for priority group calculations}

\item{...}{arguments passed to \code{\link{sqrp_level}}}
}
\value{
a data.frame with SQRP points and levels
}
\description{
\code{simulate_once_sqrp_data} produces a single iteration
of \code{\link{simulate_sqrp}}.  Essentially it is a building block function.
}
