% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/school_growth_percentile.R
\name{truncated_growth}
\alias{truncated_growth}
\title{Calculate truncated growth}
\usage{
truncated_growth(truncation_percentile = 0.99, typical_growth, sd_growth,
  start_rit = 0)
}
\arguments{
\item{truncation_percentile}{percentile to truncate to.}

\item{typical_growth}{the typical growth (expected growth) from NWEA
norms table}

\item{sd_growth}{the standard deviation growth (expected growth) from NWEA
norms table}

\item{start_rit}{the starting rit score to which truncated growth will be added.}
}
\value{
vector of RIT scores equivilent to specified growth percentile
}
\description{
\code{truncated_growth} calculates rit score associated
with a given student growth percentile
}

